refactored occurrences and occurrence / event data
