+++ EventData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bd81c4e8ddc2c493991bf2f304281634
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
