+++ OccurrenceData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e100771d9a06f43f6bbe6e2be337b7d2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
